Apply build constraints to browser tests
